Extract the calculations of AxisProjectedValues
